Refatorando linkador de recibos
