+++ lib/tasks/crawler.rake
@@ -1,4 +1,5 @@
 require "#{Rails.root}/lib/camara/camara_collector.rb"
+require "#{Rails.root}/lib/camara/recibos_collector.rb"
 
 namespace :crawler do
   desc "Baixar dados do site da camara dos deputados"
@@ -6,5 +7,6 @@ namespace :crawler do
   	 	camara_collector = CamaraCollector.new
   	 	camara_collector.recuperar_deputados
   	 	camara_collector.recuperar_cota_parlamentar
+      RecibosCollector.run
   end
 end
